requirements.txt
# APEX SURVIVOR - Python Dependencies

# Core dependencies
numpy>=1.24.0
PyYAML>=6.0

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0

# Development dependencies (optional)
# black>=23.0.0  # コードフォーマッター
# flake8>=6.0.0  # リンター
# mypy>=1.0.0    # 型チェッカー
